permisos de usuarios aplicados a todas las vistas. correccion de permisos en pagos y usurios

+++ apps/pagos/views.py
@@ -10,16 +10,21 @@ from apps.permisos.models import Permiso
 from datetime import date, datetime
 from apps.documentos.models import TipoDocumento, Documento
 from django.shortcuts import redirect
-from apps.generales.views import GenericListadoView,GenericAltaView, GenericEliminarView
+from apps.generales.views import GenericListadoView,GenericAltaView, GenericEliminarView,GenericModificacionView
 from .filters import CobrosFilter,CobrosTodosFilter, ModulosFilter, PagosFilter, PagosTodosFilter
 from .tables import CobrosTable, ModulosTable, CobrosTodosTable, PagosTable, PagosTodosTable
 from datetime import datetime
+from django.contrib.auth.mixins import LoginRequiredMixin
+from django.contrib import messages as Messages
+from django.contrib.auth.decorators import permission_required
 
 class AltaValorDeModulo(GenericAltaView):
 	model = ValorDeModulo
 	form_class = RegistrarValorDeModuloForm
 	template_name = 'pagos/modulos/alta.html'
 	success_url = reverse_lazy('pagos:listarModulos')
+	permission_required = 'pagos.cargar_valor_de_modulo'
+	redirect_url = 'pagos:listarModulos'
 
 	def get_context_data(self, **kwargs):
 		context = super(AltaValorDeModulo, self).get_context_data(**kwargs)
@@ -27,11 +32,13 @@ class AltaValorDeModulo(GenericAltaView):
 		context['ayuda'] = 'general.html#como-crear-un-nuevo-valor-de-modulo'
 		return context
 
-class ModificarValorDeModulo(UpdateView):
+class ModificarValorDeModulo(GenericModificacionView):
 	model = ValorDeModulo
 	form_class = RegistrarValorDeModuloForm
 	template_name = 'pagos/modulos/alta.html'
 	success_url = reverse_lazy('pagos:listarModulos')
+	permission_required = 'pagos.modificar_valor_de_modulo'
+	redirect_url = 'pagos:listarModulos'
 
 	def get_context_data(self, **kwargs):
 		context = super(ModificarValorDeModulo, self).get_context_data(**kwargs)
@@ -47,6 +54,8 @@ class ListadoValoresDeModulo(GenericListadoView):
 	filterset_class = ModulosFilter
 	export_name = 'listado_valores_modulos'
 	context_object_name = 'modulos'
+	permission_required = 'pagos.listar_valor_de_modulo'
+	redirect_url = '/'
 
 	def get_context_data(self, **kwargs):
 		context = super(ListadoValoresDeModulo, self).get_context_data(**kwargs)
@@ -56,19 +65,42 @@ class ListadoValoresDeModulo(GenericListadoView):
 			}
 		return context
 
-class EliminarValorDeModulo(DeleteView):
+class EliminarValorDeModulo(GenericEliminarView):
 	model = ValorDeModulo
-	template_name = 'delete.html'
-	success_url = reverse_lazy('pagos:listarModulos')
+	permission_required = 'pagos.eliminar_valor_de_modulo'
 
-class AltaCobro(GenericAltaView):
+	def post(self, request, *args, **kwargs):
+		if request.user.has_perm(self.permission_required):
+			self.object = self.get_object()
+			try:
+				self.object.delete()
+				return JsonResponse({
+					"success": True,
+					"message": 'Valor de Modulo eliminado correctamente'
+				})
+			except Exception:
+				return JsonResponse({
+				"success": False,
+				"message": ('error',"No se pudo eliminar el valor de modulo")
+				})
+		else:
+			return JsonResponse({
+					"success": False,
+					"message": ('permiso','No posee los permisos necesarios para realizar para realizar esta operación')
+			}) 
+
+class AltaCobro(LoginRequiredMixin, CreateView):
 	model = Cobro
 	form_class = CobroForm
 	template_name = 'pagos/cobros/alta.html'
 	success_url = reverse_lazy('pagos:listarCobros')
+	permission_required = 'pagos.cargar_cobro'
 
 	def get(self, request, *args, **kwargs):
 		permiso = Permiso.objects.get(pk=kwargs.get('pk'))
+		if not request.user.has_perm(self.permission_required):
+			Messages.error(self.request, 'No posee los necesarios para realizar permisos para realizar esta operación')
+			return HttpResponseRedirect(reverse('pagos:listarCobros', args=[permiso.pk]))
 		fecha = date.today() if date.today() <= permiso.fechaVencimiento else permiso.fechaVencimiento
 		cobro = permiso.estado.recalcular(usuario=request.user, documento=None, fecha=fecha, unidad=permiso.unidad) #TODO corregir fecha, el limite es HOY y el VENC. de permiso
 		return render(request, self.template_name, {'form':self.form_class(initial={'fecha_desde':cobro.fecha_desde,'fecha':cobro.fecha}), 'cobro': cobro, 
@@ -78,6 +110,9 @@ class AltaCobro(GenericAltaView):
 	def post(self, request, *args, **kwargs):
 		self.object = self.get_object
 		permiso = Permiso.objects.get(pk=kwargs.get('pk'))
+		if not request.user.has_perm(self.permission_required):
+			Messages.error(self.request, 'No posee los necesarios para realizar permisos para realizar esta operación')
+			return HttpResponseRedirect(reverse('permisos:detalle', args=[permiso.pk]))
 		documento_form = self.form_class(request.POST, request.FILES)
 
 		if documento_form.is_valid():
@@ -92,10 +127,14 @@ class AltaCobro(GenericAltaView):
 			return HttpResponseRedirect(reverse('pagos:listarCobros', args=[permiso.id]))
 		return render(request, self.template_name, {'form':documento_form, 'cobro': cobro, 'permiso': permiso})
 
+@permission_required('pagos.recalcular_cobro', login_url="/")
 def recalcular_cobro(request):
 	permiso = Permiso.objects.get(pk=request.GET['permiso_pk'])
 	fecha = datetime.strptime(request.GET['fecha'], '%Y-%m-%d').date()
-	
+	if not request.user.has_perm(self.permission_required):
+		Messages.error(self.request, 'No posee los necesarios para realizar permisos para realizar esta operación')
+		return JsonResponse({"success": False,"message": "No posee los permisos necesarios"})
+
 	if fecha <= permiso.fechaVencimiento:
 		try:
 			cobro = permiso.estado.recalcular(usuario=request.user, documento=None, fecha=fecha, unidad=permiso.unidad)
@@ -108,17 +147,22 @@ def recalcular_cobro(request):
 
 class EliminarCobro(GenericEliminarView):
 	model = Cobro
-	
+	permission_required = 'pagos.eliminar_cobro'
+
 	def post(self, request, *args, **kwargs):
 		self.object = self.get_object()
+
+		if not request.user.has_perm(self.permission_required):
+			return JsonResponse({"success": False,"message": ('permiso',"No posee los necesarios para realizar permisos para realizar esta operación")})
+
 		try:
 			if self.object == self.object.permiso.cobros.latest() and self.object.documento.tipo.slug != 'resolucion':
 				self.object.delete()
 				return JsonResponse({"success": True,"message": "Cobro eliminado"})
 			else:
-				return JsonResponse({"success": False,"message": "No se puede eliminar el primer cobro, o bien existen otros cobros posteriores al que desea eliminar"})
+				return JsonResponse({"success": False,"message": ('error',"No se puede eliminar el primer cobro, o bien existen otros cobros posteriores al que desea eliminar")})
 		except:
-			return JsonResponse({"success": False,"message": "No se ha podido realizar la eliminación del cobro"})
+			return JsonResponse({"success": False,"message": ('error',"No se ha podido realizar la eliminación del cobro")})
 
 
 class ListarCobros(GenericListadoView):
@@ -129,12 +173,16 @@ class ListarCobros(GenericListadoView):
 	filterset_class = CobrosFilter
 	export_name = 'listado_cobros'
 	context_object_name = 'cobros'
+	permission_required = 'pagos.listar_cobro'
 
 	def get_table_data(self):
 		return self.permiso.cobros.all()
 
 	def get(self, request, *args, **kwargs):
 		self.permiso = Permiso.objects.get(pk=kwargs.get('pk'))
+		if not request.user.has_perm(self.permission_required):
+			Messages.error(self.request, 'No posee los necesarios para realizar permisos para realizar esta operación')
+			return HttpResponseRedirect(reverse('permisos:detalle', args=[self.permiso.pk]))
 		return super(ListarCobros,self).get(request, *args, **kwargs)
 
 	def get_context_data(self, **kwargs):
@@ -146,13 +194,17 @@ class ListarCobros(GenericListadoView):
 		return context
 
 
-class AltaPago(GenericAltaView):
+class AltaPago(LoginRequiredMixin,CreateView):
 	model = Pago
 	form_class = DocumentoProtegidoForm
 	template_name = 'pagos/alta.html'
+	permission_required = 'pagos.cargar_pago'
 
 	def get(self, request, *args, **kwargs):
 		self.permiso_pk = kwargs.get('pk')
+		if not request.user.has_perm(self.permission_required):
+			Messages.error(self.request, 'No posee los necesarios para realizar permisos para realizar esta operación')
+			return HttpResponseRedirect(reverse('pagos:listarPagos', args=[self.permiso_pk]))
 		permiso = Permiso.objects.get(pk=kwargs.get('pk'))
 		documento_form = self.form_class()
 		documento_form.fields['fecha'].label = 'Fecha de Pago'
@@ -168,16 +220,20 @@ class AltaPago(GenericAltaView):
 		return post_pago_nuevo_modificado(self, request, documento_form, None, permiso)
 
 
-class ModificarPago(UpdateView):
+class ModificarPago(LoginRequiredMixin,UpdateView):
 	model = Pago
 	form_class = DocumentoProtegidoForm
 	template_name = 'pagos/alta.html'
 	context_object_name = 'pago'
 	success_url = reverse_lazy('pagos:listarTodosLosPagos')
+	permission_required = 'pagos.modificar_pago'
 
 	def get(self, request, *args, **kwargs):
 		super().get(request, *args, **kwargs)
 		pago = self.object
+		if not request.user.has_perm(self.permission_required):
+			Messages.error(self.request, 'No posee los necesarios para realizar permisos para realizar esta operación')
+			return HttpResponseRedirect(reverse('pagos:listarPagos', args=[kwargs.get('pkp')]))
 		form_class = self.form_class(initial={'descripcion':pago.documento.descripcion,	'archivo': pago.documento.archivo, 'fecha': pago.documento.fecha})
 		return render(request, self.template_name, { 'form': form_class,
 			'monto' : pago.monto, 'nombreForm' : 'Modificar Pago de Canon', 'return_path' : reverse('pagos:listarPagos', args=[pago.permiso.pk]),
@@ -185,6 +241,9 @@ class ModificarPago(UpdateView):
 
 	def post(self, request, *args, **kwargs):
 		super().post(request, *args, **kwargs)
+		if not request.user.has_perm(self.permission_required):
+			Messages.error(self.request, 'No posee los necesarios para realizar permisos para realizar esta operación')
+			return HttpResponseRedirect(reverse('pagos:listarPagos', args=[kwargs.get('pk')]))
 		pago = self.object
 		documento_form = self.form_class(request.POST, request.FILES)
 		documento_form.fields['archivo'].required = False
@@ -231,12 +290,16 @@ class ListarPagos(GenericListadoView):
 	filterset_class = PagosFilter
 	export_name = 'listado_pagos'
 	context_object_name = 'pagos'
+	permission_required = 'pagos.listar_pago'
 
 	def get_table_data(self):
 		return self.permiso.pagos.all()
 
 	def get(self, request, *args, **kwargs):
 		self.permiso = Permiso.objects.get(pk=kwargs.get('pk'))
+		if not request.user.has_perm(self.permission_required):
+			Messages.error(self.request, 'No posee los necesarios para realizar permisos para realizar esta operación')
+			return HttpResponseRedirect(reverse('permisos:detalle', args=[self.permiso.pk]))
 		return super(ListarPagos,self).get(request, *args, **kwargs)
 
 	def get_context_data(self, **kwargs):
@@ -249,14 +312,17 @@ class ListarPagos(GenericListadoView):
 
 class EliminarPago(GenericEliminarView):
 	model = Pago
-	
+	permission_required = 'pagos.eliminar_pago'
+
 	def post(self, request, *args, **kwargs):
+		if not request.user.has_perm(self.permission_required):
+			return JsonResponse({"success": False,"message": ('permiso',"No posee los necesarios para realizar permisos para realizar esta operación")})
 		self.object = self.get_object()
 		self.object.delete()
 		try:
 			return JsonResponse({"success": True,"message": "Pago eliminado"})
 		except:
-			return JsonResponse({"success": False,"message": "No se puede eliminar el pago de "})
+			return JsonResponse({"success": False,"message": ('error',"No se puede eliminar el pago de ")})
 
 class ListarTodosLosCobros(GenericListadoView):
 	model = Cobro
@@ -266,6 +332,8 @@ class ListarTodosLosCobros(GenericListadoView):
 	filterset_class = CobrosTodosFilter
 	export_name = 'listado_cobros_todos'
 	context_object_name = 'cobros'
+	permission_required = 'pagos.listar_todos_cobros'
+	redirect_url = '/'
 
 	def get_context_data(self, **kwargs):
 		context = super(ListarTodosLosCobros, self).get_context_data(**kwargs)
@@ -279,6 +347,8 @@ class ListarTodosLosPagos(GenericListadoView):
 	paginate_by = 12
 	filterset_class = PagosTodosFilter
 	export_name = 'listado_pagos_todos'
+	permission_required = 'pagos.listar_todos_pagos'
+	redirect_url = '/'
 
 	def get_context_data(self, **kwargs):
 		context = super(ListarTodosLosPagos, self).get_context_data(**kwargs)
@@ -286,11 +356,12 @@ class ListarTodosLosPagos(GenericListadoView):
 		return context
 
 
-class AltaCobroInfraccion(GenericAltaView):
+class AltaCobroInfraccion(LoginRequiredMixin,CreateView):
 	model = Cobro
 	form_class = DocumentoProtegidoForm
 	template_name = 'pagos/alta.html'
 	success_url = reverse_lazy('pagos:listarCobros')
+	permission_required = 'pagos.cargar_cobro_infraccion'
 
 	def get_context_data(self, **kwargs):
 		context = super(AltaCobroInfraccion, self).get_context_data(**kwargs)
@@ -302,11 +373,17 @@ class AltaCobroInfraccion(GenericAltaView):
 
 	def get (self, request, *args, **kwargs):
 		self.permiso_pk = kwargs.get('pk')
+		if not request.user.has_perm(self.permission_required):
+			Messages.error(self.request, 'No posee los necesarios para realizar permisos para realizar esta operación')
+			return HttpResponseRedirect(reverse('permisos:detalle', args=[self.permiso_pk]))
 		return super(AltaCobroInfraccion, self).get(request,*args,**kwargs)	
 
 	def post(self, request, *args, **kwargs):
 		self.object = self.get_object
 		permiso = Permiso.objects.get(pk=kwargs.get('pk'))
+		if not request.user.has_perm(self.permission_required):
+			Messages.error(self.request, 'No posee los necesarios para realizar permisos para realizar esta operación')
+			return HttpResponseRedirect(reverse('permisos:detalle', args=[permiso.pk]))
 
 		documento_form = DocumentoProtegidoForm(request.POST, request.FILES)
 		documento = documento_form.save(commit=False)
@@ -330,11 +407,12 @@ class AltaCobroInfraccion(GenericAltaView):
 		return render(request, self.template_name, {'form':documento_form, 'return_path': reverse('permisos:detalle', args=[self.permiso_pk]), 'message_error':['Error en la carga']})
 
 
-class AltaPagoInfraccion(GenericAltaView):
+class AltaPagoInfraccion(LoginRequiredMixin,CreateView):
 	model = Pago
 	form_class = DocumentoProtegidoForm
 	template_name = 'pagos/alta.html'
 	success_url = reverse_lazy('pagos:listarPagos')
+	permission_required = 'pagos.cargar_pago_infraccion'
 
 	def get_context_data(self, **kwargs):
 		context = super(AltaPagoInfraccion, self).get_context_data(**kwargs)
@@ -346,12 +424,17 @@ class AltaPagoInfraccion(GenericAltaView):
 
 	def get(self, request, *args, **kwargs):
 		self.permiso_pk = kwargs.get('pk')
+		if not request.user.has_perm(self.permission_required):
+			Messages.error(self.request, 'No posee los necesarios para realizar permisos para realizar esta operación')
+			return HttpResponseRedirect(reverse('permisos:detalle', args=[self.permiso_pk]))
 		return super(AltaPagoInfraccion, self).get(request,*args,**kwargs)
 
 	def post(self, request, *args, **kwargs):
 		self.object = self.get_object
 		permiso = Permiso.objects.get(pk=kwargs.get('pk'))
-
+		if not request.user.has_perm(self.permission_required):
+			Messages.error(self.request, 'No posee los necesarios para realizar permisos para realizar esta operación')
+			return HttpResponseRedirect(reverse('permisos:detalle', args=[permiso.pk]))
 		documento_form = self.form_class(request.POST, request.FILES)
 		monto = float(request.POST['monto'])
 		fecha_de_pago = datetime.strptime(documento_form.data['fecha'], "%Y-%m-%d").date()
