+++ apps/personas/views/empresas.py
@@ -2,15 +2,17 @@ from django.core.urlresolvers import reverse_lazy
 from apps.personas.models import Empresa
 from apps.personas.forms import *
 from django.views.generic import CreateView, ListView, DetailView, UpdateView, DeleteView
-from apps.generales.views import GenericAltaView
+from apps.generales.views import GenericAltaView,GenericDetalleView,GenericModificacionView
 from django.http import JsonResponse
 from django.core import serializers
 from django.core.serializers.json import DjangoJSONEncoder
-
+from django.contrib.auth.mixins import LoginRequiredMixin,PermissionRequiredMixin
 from django.views import View
 from django_tables2.views import SingleTableView
 from apps.personas.tables import EmpresaTable
 from django.core.urlresolvers import reverse
+from django.contrib import messages as Messages
+from django.shortcuts import redirect
 
 class AltaEmpresa(GenericAltaView):
 	model = Empresa
@@ -22,9 +24,11 @@ class AltaEmpresa(GenericAltaView):
 	permission_required = 'personas.cargar_empresa'
 	redirect_url = 'empresas:listado'
 
-class DetalleEmpresa(DetailView):
+class DetalleEmpresa(GenericDetalleView):
 	model = Empresa
 	template_name = 'empresas/detalle.html'
+	permission_required = 'personas.detalle_empresa'
+	redirect_url = 'empresas:listado'
 
 	def get_context_data(self, **kwargs):
 		context = super(DetalleEmpresa, self).get_context_data(**kwargs)
@@ -34,11 +38,17 @@ class DetalleEmpresa(DetailView):
 		context['ayuda'] = 'solicitante.html#como-crear-una-nueva-empresa'
 		return context
 
-class Listado(SingleTableView):
+class Listado(LoginRequiredMixin,PermissionRequiredMixin,SingleTableView):
 	model = Empresa
 	template_name = 'empresas/listado.html'
 	table_class = EmpresaTable
 	paginate_by = 12
+	permission_required = 'personas.listar_empresa'
+	redirect_url = '/'
+
+	def handle_no_permission(self):
+		Messages.error(self.request, 'No posee los necesarios para realizar permisos para realizar esta operación')
+		return redirect(self.redirect_url)
 
 class DataEmpresas(View):
 	def get(self, request, *args, **kwargs):
@@ -48,11 +58,14 @@ class DataEmpresas(View):
 			empresaDict.append({"razonSocial": empresa.razonSocial, "cuit": empresa.cuit})
 		return JsonResponse({"data": empresaDict})
 
-class ModificarEmpresa(UpdateView):
+class ModificarEmpresa(GenericModificacionView):
 	model = Empresa
 	form_class = EmpresaForm
 	success_url = reverse_lazy('empresas:listado')
 	template_name = 'empresas/alta.html'
+	permission_required = 'personas.modificar_empresa'
+	redirect_url = 'empresas:listado'
+	
 	def get_context_data(self, **kwargs):
 		context = super(ModificarEmpresa, self).get_context_data(**kwargs)
 		context['nombreForm'] = 'Detalle Empresa'
@@ -61,11 +74,14 @@ class ModificarEmpresa(UpdateView):
 		context['ayuda'] = 'solicitante.html#como-crear-una-nueva-empresa'
 		return context
 
-class EliminarEmpresa(DeleteView):
+class EliminarEmpresa(LoginRequiredMixin, DeleteView):
 	model = Empresa
 	success_url = reverse_lazy('empresas:listado')
+	permission_required = 'personas.eliminar_empresa'
 	
 	def delete(self, request, *args, **kwargs):
+		if not request.user.has_perm(self.permission_required):
+			return JsonResponse({"success": False,"message": ('permiso',"No posee los necesarios para realizar permisos para realizar esta operación")})
 		try:
 			self.object = self.get_object()
 			self.object.delete()
@@ -76,5 +92,5 @@ class EliminarEmpresa(DeleteView):
 		except Error as e:
 			return JsonResponse({
 				"success": False,
-				"message": e.value()
+				"message": ('error',"No se pudo eliminar la empresa")
 			})
